feat(backend): wrapping for now, added some more tweaks to the model
