+++ functions/jest.config.js
@@ -1,4 +1,9 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tests/tsconfig.json'
+    }
+  }
 };
