+++ functions/package.json
@@ -27,6 +27,7 @@
     "node": "10"
   },
   "dependencies": {
+    "@google-cloud/firestore": "^3.7.3",
     "@googlemaps/google-maps-services-js": "^2.4.2",
     "@types/googlemaps": "^3.39.3",
     "class-validator": "^0.11.1",
