+++ functions/src/models/requests/privilegedInformation.ts
@@ -1,0 +1,6 @@
+import { firestore } from 'firebase-admin';
+import DocumentData = firestore.DocumentData;
+
+export interface PrivilegedRequestInformation extends DocumentData {
+  address: google.maps.GeocoderResult;
+}
