+++ functions/src/models/users/privilegedInformation.ts
@@ -1,0 +1,10 @@
+import { firestore } from 'firebase-admin';
+import Timestamp = firestore.Timestamp;
+import DocumentData = firestore.DocumentData;
+
+export interface PrivilegedUserInformation extends DocumentData {
+  address: google.maps.GeocoderResult;
+
+  termsAccepted: Timestamp; // acts as a timestamp of when and as a boolean: if accepted it exists.
+  termsVersion: string;
+}
