+++ functions/src/requests/priviledgedInformation.ts
@@ -1,9 +1,5 @@
 import * as functions from 'firebase-functions';
 
-export interface PrivateRequestData extends FirebaseFirestore.DocumentData {
-  address: google.maps.GeocoderResult;
-}
-
 export const triggerEventsWhenRequestPrivilegedInformationIsCreated =
   functions.firestore.document('requests/{requestId}/privilegedInformation/{informationId}')
     // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
