+++ functions/tests/index.test.ts
@@ -1,10 +1,12 @@
 import { validate } from 'class-validator';
 
-import { User } from '../src/users';
+import { User } from '../src/models/users';
 
 describe('user validation', () => {
   it('fails when you pass it negative ratings ', () => {
     const user = User.factory({
+        cavQuestionnaireRef: null,
+        pinQuestionnaireRef: null,
         averageRating: -1,
         casesCompleted: 0,
         requestsMade: 0,
