+++ functions/tests/tsconfig.json
@@ -12,6 +12,6 @@
   },
   "compileOnSave": true,
   "include": [
-    "."
+    "./"
   ]
 }
